requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
motor==3.4.0
multidict==6.0.5
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
yarl==1.9.4
python-dotenv
slowapi
pytz